Move try_connect_inner to connect.rs
